requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
alembic==1.5.8
aniso8601==7.0.0
async-timeout==3.0.1
asyncpg==0.22.0
asyncpgsa==0.27.1
attrs==20.3.0
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.0.0
h11==0.12.0
idna==2.10
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
promise==2.3
psycopg2-binary==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
yarl==1.6.3
